Move components and utils to packages: * TruncatedText and TruncatedTextWithTooltip to @geti/ui * idMatchingFormat to @geti/ui/utils
