+++ web_ui/package.json
@@ -128,6 +128,7 @@
         "@types/adm-zip": "^0.5.7",
         "@types/jest": "^29.5.14",
         "@types/lodash-es": "^4.17.12",
+        "@types/mocha": "^10.0.10",
         "@types/node": "^22.15.3",
         "@types/polylabel": "^1.1.3",
         "@types/react": "^19.1.10",
