+++ web_ui/packages/ui/index.ts
@@ -161,6 +161,7 @@ export { MediaViewModes } from './src/view-modes/media-view-modes.component';
 export { ViewModes, INITIAL_VIEW_MODE, VIEW_MODE_LABEL } from './src/view-modes/utils';
 export { useViewMode } from './src/view-modes/use-view-mode.hook';
 export { Toast, toast, removeToasts, removeToast, CustomToast } from './src/toast/toast.component';
+export { TruncatedText, TruncatedTextWithTooltip } from './src/truncated-text/truncated-text.component';
 
 export {
     ListBox as AriaComponentsListBox,
