+++ web_ui/packages/ui/package.json
@@ -1,44 +1,45 @@
 {
-  "name": "@geti/ui",
-  "version": "1.0.0",
-  "description": "Contains UI representative components",
-  "main": "index.ts",
-  "type": "module",
-  "dependencies": {
-    "@adobe/react-spectrum": "^3.43.0",
-    "@spectrum-icons/illustrations": "^3.6.24",
-    "@spectrum-icons/workflow": "^4.2.23",
-    "lodash-es": "^4.17.21",
-    "react-aria-components": "^1.11.0",
-    "react-colorful": "^5.6.1",
-    "sonner": "^2.0.7",
-    "usehooks-ts": "^3.1.1"
-  },
-  "devDependencies": {
-    "@geti/config": "*",
-    "@testing-library/dom": "^10.4.1",
-    "@testing-library/jest-dom": "^6.6.4",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
-    "@types/lodash-es": "^4.17.12",
-    "eslint": "^9.25.1",
-    "eslint-plugin-header": "^3.1.1",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-playwright": "^2.2.0",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "typescript": "^5.8.3"
-  },
-  "scripts": {
-    "test:unit": "npm run test:unit --workspace=root --config jest.config.js --watchAll=false",
-    "type-check": "tsc --noEmit -p tsconfig.json",
-    "lint": "eslint \"src/**/*.ts*\" --max-warnings 0",
-    "lint:fix": "eslint --fix \"./src/**/*.ts*\""
-  },
-  "exports": {
-    ".": "./index.ts",
-    "./icons": "./icons/index.ts",
-    "./theme": "./theme/index.tsx"
-  }
+    "name": "@geti/ui",
+    "version": "1.0.0",
+    "description": "Contains UI representative components",
+    "main": "index.ts",
+    "type": "module",
+    "dependencies": {
+        "@adobe/react-spectrum": "^3.43.0",
+        "@spectrum-icons/illustrations": "^3.6.24",
+        "@spectrum-icons/workflow": "^4.2.23",
+        "lodash-es": "^4.17.21",
+        "react-aria-components": "^1.11.0",
+        "react-colorful": "^5.6.1",
+        "sonner": "^2.0.7",
+        "usehooks-ts": "^3.1.1"
+    },
+    "devDependencies": {
+        "@geti/config": "*",
+        "@testing-library/dom": "^10.4.1",
+        "@testing-library/jest-dom": "^6.6.4",
+        "@testing-library/react": "^16.3.0",
+        "@testing-library/user-event": "^14.6.1",
+        "@types/lodash-es": "^4.17.12",
+        "eslint": "^9.25.1",
+        "eslint-plugin-header": "^3.1.1",
+        "eslint-plugin-import": "^2.32.0",
+        "eslint-plugin-jsx-a11y": "^6.10.2",
+        "eslint-plugin-playwright": "^2.2.0",
+        "eslint-plugin-react": "^7.37.5",
+        "eslint-plugin-react-hooks": "^5.2.0",
+        "typescript": "^5.8.3"
+    },
+    "scripts": {
+        "test:unit": "npm run test:unit --workspace=root --config jest.config.js --watchAll=false",
+        "type-check": "tsc --noEmit -p tsconfig.json",
+        "lint": "eslint \"src/**/*.ts*\" --max-warnings 0",
+        "lint:fix": "eslint --fix \"./src/**/*.ts*\""
+    },
+    "exports": {
+        ".": "./index.ts",
+        "./icons": "./icons/index.ts",
+        "./theme": "./theme/index.tsx",
+        "./utils": "./utils/index.ts"
+    }
 }
